split index logic in zettels and links
