+++ aftermath/pkg/index/index_test.go
@@ -101,7 +101,7 @@ func TestUpdateZettelIndex(t *testing.T) {
 }
 
 // TestIndexLinks tests the indexing of links between zettels.
-func TestIndexLinks(t *testing.T) {
+func TestUpdateLinkIndex(t *testing.T) {
 	db := openTestDB(t)
 	defer closeTestDB(t, db)
 
@@ -179,27 +179,6 @@ func TestIndexLinks(t *testing.T) {
 			links[0].SourceID,
 			links[0].TargetID,
 		)
-	}
-}
-
-func TestLink2File(t *testing.T) {
-	testIndexer := index.NewIndexer(nil, "/base/dir")
-	tests := []struct {
-		input    string
-		expected string
-	}{
-		{"foo.bar", "/base/dir/foo/bar.typ"},
-		{"foo.bar.baz", "/base/dir/foo/bar/baz.typ"},
-		{"foo_bar", "/base/dir/foo_bar.typ"},
-		{"foo-bar", "/base/dir/foo-bar.typ"},
-		{"foo.bar-baz.qux", "/base/dir/foo/bar-baz/qux.typ"},
-	}
-
-	for _, test := range tests {
-		result := testIndexer.Link2File(test.input)
-		if result != test.expected {
-			t.Errorf("link2File(%q) = %q; expected %q", test.input, result, test.expected)
-		}
 	}
 }
